backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
minio==7.2.7
Pillow==10.0.0
gunicorn==21.2.0